==> code/01_draft_results.ipynb <==
## Draft Results

This notebook combines and cleans all draft datasets into a single dataset containing all drafted players between 2001 and 2022.

import pandas as pd
import numpy as np

draft01 = pd.read_csv('../data/drafts/2001_Draft.csv')
draft02 = pd.read_csv('../data/drafts/2002_Draft.csv')
draft03 = pd.read_csv('../data/drafts/2003_Draft.csv')
draft04 = pd.read_csv('../data/drafts/2004_Draft.csv')
draft05 = pd.read_csv('../data/drafts/2005_Draft.csv')
draft06 = pd.read_csv('../data/drafts/2006_Draft.csv')
draft07 = pd.read_csv('../data/drafts/2007_Draft.csv')
draft08 = pd.read_csv('../data/drafts/2008_Draft.csv')
draft09 = pd.read_csv('../data/drafts/2009_Draft.csv')
draft10 = pd.read_csv('../data/drafts/2010_Draft.csv')
draft11 = pd.read_csv('../data/drafts/2011_Draft.csv')
draft12 = pd.read_csv('../data/drafts/2012_Draft.csv')
draft13 = pd.read_csv('../data/drafts/2013_Draft.csv')
draft14 = pd.read_csv('../data/drafts/2014_Draft.csv')
draft15 = pd.read_csv('../data/drafts/2015_Draft.csv')
draft16 = pd.read_csv('../data/drafts/2016_Draft.csv')
draft17 = pd.read_csv('../data/drafts/2017_Draft.csv')
draft18 = pd.read_csv('../data/drafts/2018_Draft.csv')
draft19 = pd.read_csv('../data/drafts/2019_Draft.csv')
draft20 = pd.read_csv('../data/drafts/2020_Draft.csv')
draft21 = pd.read_csv('../data/drafts/2021_Draft.csv')
draft22 = pd.read_csv('../data/drafts/2022_Draft.csv')
draft01 = draft01[['Rnd', 'Pick', 'Pos', '-9999']]
draft01.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft01['DraftYear'] = 2001
draft02 = draft02[['Rnd', 'Pick', 'Pos', '-9999']]
draft02.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft02['DraftYear'] = 2002
draft03 = draft03[['Rnd', 'Pick', 'Pos', '-9999']]
draft03.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft03['DraftYear'] = 2003
draft04 = draft04[['Rnd', 'Pick', 'Pos', '-9999']]
draft04.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft04['DraftYear'] = 2004
draft05 = draft05[['Rnd', 'Pick', 'Pos', '-9999']]
draft05.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft05['DraftYear'] = 2005
draft06 = draft06[['Rnd', 'Pick', 'Pos', '-9999']]
draft06.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft06['DraftYear'] = 2006
draft07 = draft07[['Rnd', 'Pick', 'Pos', '-9999']]
draft07.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft07['DraftYear'] = 2007
draft08 = draft08[['Rnd', 'Pick', 'Pos', '-9999']]
draft08.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft08['DraftYear'] = 2008
draft09 = draft09[['Rnd', 'Pick', 'Pos', '-9999']]
draft09.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft09['DraftYear'] = 2009
draft10 = draft10[['Rnd', 'Pick', 'Pos', '-9999']]
draft10.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft10['DraftYear'] = 2010
draft11 = draft11[['Rnd', 'Pick', 'Pos', '-9999']]
draft11.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft11['DraftYear'] = 2011
draft12 = draft12[['Rnd', 'Pick', 'Pos', '-9999']]
draft12.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft12['DraftYear'] = 2012
draft13 = draft13[['Rnd', 'Pick', 'Pos', '-9999']]
draft13.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft13['DraftYear'] = 2013
draft14 = draft14[['Rnd', 'Pick', 'Pos', '-9999']]
draft14.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft14['DraftYear'] = 2014
draft15 = draft15[['Rnd', 'Pick', 'Pos', '-9999']]
draft15.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft15['DraftYear'] = 2015
draft16 = draft16[['Rnd', 'Pick', 'Pos', '-9999']]
draft16.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft16['DraftYear'] = 2016
draft17 = draft17[['Rnd', 'Pick', 'Pos', '-9999']]
draft17.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft17['DraftYear'] = 2017
draft18 = draft18[['Rnd', 'Pick', 'Pos', '-9999']]
draft18.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft18['DraftYear'] = 2018
draft19 = draft19[['Rnd', 'Pick', 'Pos', '-9999']]
draft19.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft19['DraftYear'] = 2019
draft20 = draft20[['Rnd', 'Pick', 'Pos', '-9999']]
draft20.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft20['DraftYear'] = 2020
draft21 = draft21[['Rnd', 'Pick', 'Pos', '-9999']]
draft21.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft21['DraftYear'] = 2021
draft22 = draft22[['Rnd', 'Pick', 'Pos', '-9999']]
draft22.rename(columns = {'-9999': 'Player-additional'}, inplace = True)
draft22['DraftYear'] = 2022

draft_final = pd.concat([draft01, draft02, draft03, draft04, draft05, draft06, draft07, draft08, draft09, draft10, 
                         draft11, draft12, draft13, draft14, draft15, draft16, draft17, draft18, draft19, draft20, 
                         draft21, draft22])
draft_final

draft_final.to_csv('../data/draft_final.csv', index = False)
