==> code/02_debut_year.ipynb <==
## Draft Results

This notebook combines and cleans all debut datasets into a single dataset containing all players who made their first NFL start between 2001 and 2021.

import pandas as pd
import numpy as np

debut01 = pd.read_csv('../data/debuts/2001_Debuts.csv')
debut02 = pd.read_csv('../data/debuts/2002_Debuts.csv')
debut03 = pd.read_csv('../data/debuts/2003_Debuts.csv')
debut04 = pd.read_csv('../data/debuts/2004_Debuts.csv')
debut05 = pd.read_csv('../data/debuts/2005_Debuts.csv')
debut06 = pd.read_csv('../data/debuts/2006_Debuts.csv')
debut07 = pd.read_csv('../data/debuts/2007_Debuts.csv')
debut08 = pd.read_csv('../data/debuts/2008_Debuts.csv')
debut09 = pd.read_csv('../data/debuts/2009_Debuts.csv')
debut10 = pd.read_csv('../data/debuts/2010_Debuts.csv')
debut11 = pd.read_csv('../data/debuts/2011_Debuts.csv')
debut12 = pd.read_csv('../data/debuts/2012_Debuts.csv')
debut13 = pd.read_csv('../data/debuts/2013_Debuts.csv')
debut14 = pd.read_csv('../data/debuts/2014_Debuts.csv')
debut15 = pd.read_csv('../data/debuts/2015_Debuts.csv')
debut16 = pd.read_csv('../data/debuts/2016_Debuts.csv')
debut17 = pd.read_csv('../data/debuts/2017_Debuts.csv')
debut18 = pd.read_csv('../data/debuts/2018_Debuts.csv')
debut19 = pd.read_csv('../data/debuts/2019_Debuts.csv')
debut20 = pd.read_csv('../data/debuts/2020_Debuts.csv')
debut21 = pd.read_csv('../data/debuts/2021_Debuts.csv')
debut01 = debut01[['Pos', '-9999']]
debut01.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut01['RookieYear'] = 2001
debut02 = debut02[['Pos', '-9999']]
debut02.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut02['RookieYear'] = 2002
debut03 = debut03[['Pos', '-9999']]
debut03.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut03['RookieYear'] = 2003
debut04 = debut04[['Pos', '-9999']]
debut04.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut04['RookieYear'] = 2004
debut05 = debut05[['Pos', '-9999']]
debut05.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut05['RookieYear'] = 2005
debut06 = debut06[['Pos', '-9999']]
debut06.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut06['RookieYear'] = 2006
debut07 = debut07[['Pos', '-9999']]
debut07.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut07['RookieYear'] = 2007
debut08 = debut08[['Pos', '-9999']]
debut08.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut08['RookieYear'] = 2008
debut09 = debut09[['Pos', '-9999']]
debut09.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut09['RookieYear'] = 2009
debut10 = debut10[['Pos', '-9999']]
debut10.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut10['RookieYear'] = 2010
debut11 = debut11[['Pos', '-9999']]
debut11.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut11['RookieYear'] = 2011
debut12 = debut12[['Pos', '-9999']]
debut12.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut12['RookieYear'] = 2012
debut13 = debut13[['Pos', '-9999']]
debut13.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut13['RookieYear'] = 2013
debut14 = debut14[['Pos', '-9999']]
debut14.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut14['RookieYear'] = 2014
debut15 = debut15[['Pos', '-9999']]
debut15.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut15['RookieYear'] = 2015
debut16 = debut16[['Pos', '-9999']]
debut16.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut16['RookieYear'] = 2016
debut17 = debut17[['Pos', '-9999']]
debut17.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut17['RookieYear'] = 2017
debut18 = debut18[['Pos', '-9999']]
debut18.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut18['RookieYear'] = 2018
debut19 = debut19[['Pos', '-9999']]
debut19.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut19['RookieYear'] = 2019
debut20 = debut20[['Pos', '-9999']]
debut20.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut20['RookieYear'] = 2020
debut21 = debut21[['Pos', '-9999']]
debut21.rename(columns = {'Pos': 'RPos', '-9999': 'Player-additional'}, inplace = True)
debut21['RookieYear'] = 2021

debut_final = pd.concat([debut01, debut02, debut03, debut04, debut05, debut06, debut07, debut08, debut09, debut10, 
                         debut11, debut12, debut13, debut14, debut15, debut16, debut17, debut18, debut19, debut20, 
                         debut21])
debut_final

debut_final.to_csv('../data/debut_final.csv', index = False)
