==> code/03_rookies.ipynb <==
## Rookies

This notebook combines the receiving dataset with the draft and debut dataset in order to get average stats of Rookies based on what round they were drafted in.

import pandas as pd
import numpy as np

def rookies(year):
    df = pd.read_csv(('../data/receiving/' + str(year) + '_receiving_basic_original.csv'))
    draft = pd.read_csv('../data/draft_final.csv')
    debut = pd.read_csv('../data/debut_final.csv')

    draft.rename(columns = {'Pos': 'DPos'}, inplace = True)

    df = df.merge(draft, how = 'left', on = 'Player-additional')
    df = df.merge(debut, how = 'left', on = 'Player-additional')

    df.drop(columns = ['Rk', 'Pos', 'DPos', 'DraftYear'], inplace = True)

    df['Rnd'].fillna(8, inplace = True)
    df['Pick'].fillna(260, inplace = True)
    
    df = df[df['RookieYear'] == year]

    return df

df1 = rookies(2021)
df2 = rookies(2020)
df3 = rookies(2019)
df4 = rookies(2018)
df5 = rookies(2017)
df6 = rookies(2016)

df_final = pd.concat([df1, df2, df3, df4, df5, df6])
df_final = df_final[(df_final['RPos'] == 'RB') | (df_final['RPos'] == 'WR') | (df_final['RPos'] == 'TE')]

# Takes total number of Stat, and divides by total games to account for difference between 16/17 game seasons.
df_final['Y/G * 17'] = (df_final['Y/G'] * 17)
df_final['Tgt/G * 17'] = ((df_final['Tgt'] / df_final['G']) * 17)
df_final['Rec/G * 17'] = ((df_final['Rec'] / df_final['G']) * 17)
df_final['TD/G * 17'] = ((df_final['TD'] / df_final['G']) * 17)
df_final

round(df_final.groupby(['Rnd', 'RPos'])[['G', 'GS', 'Tgt', 'Rec', 'Yds', 'TD', 'Tgt/G * 17', 'Rec/G * 17', 'Y/G * 17', 'TD/G * 17']].mean(),0)

rookie_stats = pd.DataFrame(round(df_final.groupby(['Rnd', 'RPos'])[['G', 'GS', 'Tgt', 'Rec', 'Yds', 'TD',
                                                                     'Tgt/G * 17', 'Rec/G * 17', 'Y/G * 17', 'TD/G * 17']].mean(),0))
rookie_stats.to_csv('../data/rookie_stats.csv')
