==> code/08_modeling_TDs.ipynb <==
## Modeling Touchdowns

This notebook builds out a model to predict touchdowns.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV
from sklearn.preprocessing import StandardScaler, PolynomialFeatures
from sklearn.linear_model import LinearRegression, Ridge, RidgeCV, Lasso, LassoCV, ElasticNet, ElasticNetCV
from sklearn.ensemble import RandomForestRegressor, ExtraTreesRegressor 
from sklearn.ensemble import GradientBoostingRegressor, AdaBoostRegressor, BaggingRegressor, StackingRegressor
from sklearn.tree import DecisionTreeRegressor
from sklearn.neighbors import KNeighborsRegressor
from sklearn.naive_bayes import MultinomialNB
from sklearn.pipeline import Pipeline
from sklearn.decomposition import PCA
from sklearn import metrics

np.random.seed(7)

df = pd.read_csv('../data/receiving_train.csv', index_col = 'Player')
df.dropna(inplace = True)
df.head()

df.shape

### Baseline Score

df[['Tgt_target', 'Rec_target', 'Yds_target', 'TD_target']].mean()

## Models

X = df.drop(columns = ['Tm', 'Pos', 'Player-additional', 'Tgt_target', 'Rec_target', 'Yds_target', 'TD_target', 'Year'])
y = df['TD_target']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 7)

lr = LinearRegression()

lr.fit(X_train, y_train)

print(f'Training Score: {lr.score(X_train, y_train)}')
print(f'Testing Score: {lr.score(X_test, y_test)}')

knn = KNeighborsRegressor()

knn.fit(X_train, y_train)

print(f'Training Score: {knn.score(X_train, y_train)}')
print(f'Testing Score: {knn.score(X_test, y_test)}')

dt = DecisionTreeRegressor()

dt.fit(X_train, y_train)

print(f'Training Score: {dt.score(X_train, y_train)}')
print(f'Testing Score: {dt.score(X_test, y_test)}')

rf = RandomForestRegressor()

rf.fit(X_train, y_train)

print(f'Training Score: {rf.score(X_train, y_train)}')
print(f'Testing Score: {rf.score(X_test, y_test)}')

gr = GradientBoostingRegressor()

gr.fit(X_train, y_train)

print(f'Training Score: {gr.score(X_train, y_train)}')
print(f'Testing Score: {gr.score(X_test, y_test)}')

ada = AdaBoostRegressor()

ada.fit(X_train, y_train)

print(f'Training Score: {ada.score(X_train, y_train)}')
print(f'Testing Score: {ada.score(X_test, y_test)}')

## Gridsearching

I want to add some Gridsearching to see if we can improve any of our models:

rf2 = RandomForestRegressor()

rf_params = {
    'n_estimators': [100, 150, 200],
    'max_depth': [None, 1, 2, 3, 4, 5]
}
gs = GridSearchCV(rf2, param_grid = rf_params, cv = 5)
gs.fit(X_train, y_train)
print(gs.best_score_)
print(gs.best_params_)

print(f'Training Score: {gs.score(X_train, y_train)}')
print(f'Testing Score: {gs.score(X_test, y_test)}')

gb2 = GradientBoostingRegressor()

gb_params = {
    'learning_rate': [.1, .05, .2],
    'n_estimators': [100, 150, 200],
    'max_depth': [None, 1, 2, 3, 4, 5]
}
gs2 = GridSearchCV(gb2, param_grid = gb_params, cv = 5)
gs2.fit(X_train, y_train)
print(gs2.best_score_)
print(gs2.best_params_)

print(f'Training Score: {gs2.score(X_train, y_train)}')
print(f'Testing Score: {gs2.score(X_test, y_test)}')

ada2 = AdaBoostRegressor()

ada_params = {
    'learning_rate': [.1, .05, .2],
    'n_estimators': [100, 150, 200],
}
gs3 = GridSearchCV(ada2, param_grid = ada_params, cv = 5)
gs3.fit(X_train, y_train)
print(gs3.best_score_)
print(gs3.best_params_)

print(f'Training Score: {gs3.score(X_train, y_train)}')
print(f'Testing Score: {gs3.score(X_test, y_test)}')

All 3 of our models look better after running the Gridsearch, but the RandomForest still looks the best. It is a little overfit, but all 3 of our models are overfit.

## Ridge

Next, I want to add regularization using a Ridge model.

X = df.drop(columns = ['Tm', 'Pos', 'Player-additional', 'Tgt_target', 'Rec_target', 'Yds_target', 'TD_target', 'Year'])
y = df['TD_target']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 7)

ss = StandardScaler()
X_train_sc = ss.fit_transform(X_train)
X_test_sc = ss.transform(X_test)

Ridge with Default alpha = 1.0

ridge = Ridge()

ridge.fit(X_train_sc, y_train)

print(f'Training Score: {ridge.score(X_train_sc, y_train)}')
print(f'Testing Score: {ridge.score(X_test_sc, y_test)}')

Find the best alpha

alphas = np.logspace(0, 5, 100)
ridge_cv = RidgeCV(alphas = alphas, cv = 5)
ridge_cv.fit(X_train_sc, y_train)

ridge_cv.alpha_

ridge_cv.best_score_

print(f'Training Score: {ridge_cv.score(X_train_sc, y_train)}')
print(f'Testing Score: {ridge_cv.score(X_test_sc, y_test)}')

## Lasso

After running a Ridge model, I ran a Lasso Model.

lasso = Lasso()

lasso.fit(X_train_sc, y_train)

print(f'Training Score: {lasso.score(X_train_sc, y_train)}')
print(f'Testing Score: {lasso.score(X_test_sc, y_test)}')

l_alphas = np.logspace(-3, 0, 100)
lasso_cv = LassoCV(alphas = l_alphas, max_iter = 10000)
lasso_cv.fit(X_train_sc, y_train)

lasso_cv.alpha_

print(f'Training Score: {lasso_cv.score(X_train_sc, y_train)}')
print(f'Testing Score: {lasso_cv.score(X_test_sc, y_test)}')

lasso_coefs = lasso_cv.coef_
lasso_coefs

lasso_coefs = pd.Series(lasso_coefs, X.columns)

lasso_coefs[lasso_coefs != 0].sort_values(ascending = False)

One Standard Deviation increase in Yds will increase target yard projections by 141.745

## ElasticNet

Because I ran both a Ridge and a Lasso model, I wanted to also run an ElasticNet, which combines the previous two models to hopefully form a better model. 

enet_alphas = np.linspace(0.5, 1.0, 100)
enet_ratio = 0.5
enet_model = ElasticNetCV(alphas = enet_alphas, l1_ratio = enet_ratio, cv=5, max_iter = 100000)
enet_model = enet_model.fit(X_train, y_train)

print(f'Training Score: {enet_model.score(X_train, y_train)}')
print(f'Testing Score: {enet_model.score(X_test, y_test)}')

Afterwards, I tried adding Polynomial features to both the Ridge, Lasso, and ElasticNet models, but it either made them worse, or threw a bunch of conversion errors, so I decided against using them.

## Principal Component Analysis

X = df.drop(columns = ['Tm', 'Pos', 'Player-additional', 'Tgt_target', 'Rec_target', 'Yds_target', 'TD_target', 'Year'])
y = df['TD_target']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 7)

ss = StandardScaler()
X_train_sc = ss.fit_transform(X_train)
X_test_sc = ss.transform(X_test)

pca = PCA(n_components = 28, random_state = 7)

X_train_pca = pca.fit_transform(X_train_sc)
X_test_pca = pca.transform(X_test_sc)

lr2 = LinearRegression()

lr2.fit(X_train_pca, y_train)

var_exp = pca.explained_variance_ratio_
print(f'Explained variance (first 20 components): {np.round(var_exp[:20], 3)}')
cum_var_exp = np.cumsum(var_exp)
print(f'Cumulative explained variance (first 20 components): {np.round(cum_var_exp[:20], 3)}')

print(f'Training Score: {round(lr2.score(X_train_pca, y_train),4)}')
print(f'Testing Score: {round(lr2.score(X_test_pca, y_test),4)}')

When running the PCA model, I was playing around with the number of components to use, and ended up finding my best result using the first 28 components.

## Ensemble Model

Now that we've tried a number of models, I will put some of our better scoring models together in a stacking model and see if we can get an even better score.

level1_models = [
    ('rf', RandomForestRegressor()),
    ('gb', GradientBoostingRegressor()),
    ('ada', AdaBoostRegressor()),
    ('lasso_pipe', Pipeline([
        ('ss', StandardScaler()),
        ('lasso', LassoCV())
    ]))
]

stack = StackingRegressor(estimators = level1_models,
                          final_estimator = LinearRegression())

cross_val_score(stack, X_train, y_train).mean()

stack.fit(X_train, y_train)

print(f'Training Score: {stack.score(X_train, y_train)}')
print(f'Testing Score: {stack.score(X_test, y_test)}')

Similar to our receptions, it appears that our RandomForest model ended up being the highest scoring model, but the stacking model is a lot less overfit than the RandomForest model, so we'll use this for our Touchdown predictions model.
