==> code/09_predictions.ipynb <==
# Final Models and Predictions

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV
from sklearn.preprocessing import StandardScaler, PolynomialFeatures
from sklearn.linear_model import LinearRegression, Ridge, RidgeCV, Lasso, LassoCV, ElasticNet, ElasticNetCV
from sklearn.ensemble import RandomForestRegressor, ExtraTreesRegressor 
from sklearn.ensemble import GradientBoostingRegressor, AdaBoostRegressor, BaggingRegressor, StackingRegressor
from sklearn.tree import DecisionTreeRegressor
from sklearn.neighbors import KNeighborsRegressor
from sklearn.naive_bayes import MultinomialNB
from sklearn.pipeline import Pipeline
from sklearn.decomposition import PCA
from sklearn import metrics

np.random.seed(7)

df_train = pd.read_csv('../data/receiving_train.csv', index_col = 'Player')
df_test = pd.read_csv('../data/receiving_test.csv', index_col = 'Player')
df_train.dropna(inplace = True)
df_test.dropna(inplace = True)

## Yardage Model

For my final Yardage model, I used the Linear Regression Model. On our training set, the model had a training score of .579, and a testing score of .533. I chose to use this model because it was not as overfit as most of our other models (such as the RandomForest), but it still had a decent testing score.

X = df_train.drop(columns = ['Tm', 'Pos', 'Player-additional', 'Tgt_target', 'Rec_target', 'Yds_target', 'TD_target', 'Year'])
y = df_train['Yds_target']

lr = LinearRegression()

lr.fit(X, y)

print(f'Training Score: {lr.score(X, y)}')

preds = lr.predict(df_test.drop(columns = ['Tm', 'Pos', 'Player-additional', 'Year']))

yard_preds = pd.DataFrame(preds, index = df_test.index).sort_values(by = 0, ascending = False)
yard_preds[0] = round(yard_preds[0])
yard_preds.rename(columns = {0: 'Yards'}, inplace = True)
yard_preds.head(20)

## Touchdowns Model

For my final Touchdown model, I used the Stacking Model that combined, RandomForest, GradientBoost, AdaBoost, and Lasso, and then fed those result into a Linear Regression model. On our training set, we had a training score of .332, and a testing score of .314. All of our touchdown models had relatively low testing scores, but I chose this model because of it’s close training and testing scores, and it was one of the higher scoring models.

X = df_train.drop(columns = ['Tm', 'Pos', 'Player-additional', 'Tgt_target', 'Rec_target', 'Yds_target', 'TD_target', 'Year'])
y = df_train['TD_target']

level1_models = [
    ('rf', RandomForestRegressor()),
    ('gb', GradientBoostingRegressor()),
    ('ada', AdaBoostRegressor()),
    ('lasso_pipe', Pipeline([
        ('ss', StandardScaler()),
        ('lasso', LassoCV())
    ]))
]

stack = StackingRegressor(estimators = level1_models,
                          final_estimator = LinearRegression())

stack.fit(X, y)

print(f'Training Score: {stack.score(X, y)}')

preds = stack.predict(df_test.drop(columns = ['Tm', 'Pos', 'Player-additional', 'Year']))

TD_preds = pd.DataFrame(preds, index = df_test.index).sort_values(by = 0, ascending = False)
TD_preds[0] = round(TD_preds[0])
TD_preds.rename(columns = {0: 'Touchdowns'}, inplace = True)
TD_preds.head(20)

## Receptions Model

For my final Receptions model, I used the Linear Regression Model. On our training set, we had a training score of .542, and a testing score of .474. I chose this model due to it’s relatively high testing score in comparison with other models, but also did not appear as overfit as some better scoring models. 

X = df_train.drop(columns = ['Tm', 'Pos', 'Player-additional', 'Tgt_target', 'Rec_target', 'Yds_target', 'TD_target', 'Year'])
y = df_train['Rec_target']

lr = LinearRegression()

lr.fit(X, y)

print(f'Training Score: {lr.score(X, y)}')

preds = lr.predict(df_test.drop(columns = ['Tm', 'Pos', 'Player-additional', 'Year']))

rec_preds = pd.DataFrame(preds, index = df_test.index).sort_values(by = 0, ascending = False)
rec_preds[0] = round(rec_preds[0])
rec_preds.rename(columns = {0: 'Receptions'}, inplace = True)
rec_preds.head(20)

## Final Predictions

The code below combines our 3 different models into a single chart and then exports it as `predictions.csv` in the data folder. Below are the top 20 players in our projection, sorted by Receptions.

final_preds = rec_preds.merge(yard_preds, how = 'left', left_index = True, right_index = True)
final_preds = final_preds.merge(TD_preds, how = 'left', left_index = True, right_index = True)
final_preds.head(20)

#final_preds.to_csv('../data/predictions.csv', index = True)
